# Archivos/openapiCliente.yaml
openapi: 3.0.3
info:
  title: Microservicio de Clientes
  description: API para la gestion de clientes en el banco XYZ.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /clientes:
    get:
      summary: Obtener todos los clientes
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
    post:
      summary: Crear un nuevo cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
      responses:
        '201':
          description: Cliente creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'

  /clientes/{id}:
    get:
      summary: Obtener cliente por ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID del cliente
          schema:
            type: integer
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Cliente no encontrado
    put:
      summary: Actualizar cliente
      parameters:
        - name: id
          in: path
          required: true
          description: ID del cliente
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
      responses:
        '200':
          description: Cliente actualizado exitosamente
        '404':
          description: Cliente no encontrado
    delete:
      summary: Eliminar cliente
      parameters:
        - name: id
          in: path
          required: true
          description: ID del cliente
          schema:
            type: integer
      responses:
        '204':
          description: Cliente eliminado exitosamente
        '404':
          description: Cliente no encontrado

components:
  schemas:
    Cliente:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        apellido:
          type: string
        dni:
          type: string
        email:
          type: string
      required:
        - nombre
        - apellido
        - dni
        - email